moved custom map implementations to its own package
