+++ rv-predict-logging/src/main/java/com/runtimeverification/rvpredict/trace/maps/MemoryAddrToValueMap.java
@@ -1,4 +1,4 @@
-package com.runtimeverification.rvpredict.trace;
+package com.runtimeverification.rvpredict.trace.maps;
 
 import java.util.Arrays;
