+++ rv-predict-logging/src/main/java/com/runtimeverification/rvpredict/trace/maps/ThreadIDToObjectMap.java
@@ -1,6 +1,8 @@
-package com.runtimeverification.rvpredict.trace;
+package com.runtimeverification.rvpredict.trace.maps;
 
 import java.util.function.Supplier;
+
+import com.runtimeverification.rvpredict.trace.LongToObjectMap;
 
 public final class ThreadIDToObjectMap<T> extends LongToObjectMap<T> {
